Nix/Darwin: Attempt to fix borgmatic timeout
Use ExitTimeOut, not TimeOut. This changes the property
name. It remains to be seen whether this works.

From `man launchd.plist`:

TimeOut <integer>
The recommended idle time out (in seconds) to pass to the job. This key
never did anything interesting and is no longer implemented. Jobs
seeking to exit when idle should use the EnablePressuredExit key to opt
into the system mechanism for reclaiming killable jobs under memory
pressure.

ExitTimeOut <integer>
The amount of time launchd waits between sending the SIGTERM signal and
before sending a SIGKILL signal when the job is to be stopped. The
default value is system-defined. The value zero is interpreted as
infinity and should not be used, as it can stall system shutdown
forever.